Refactor project.py functions
