+++ project_generator/util.py
@@ -18,3 +18,24 @@ import shutil
 def rmtree_if_exists(directory):
     if os.path.exists(directory):
         shutil.rmtree(directory)
+
+def merge_recursive(*args):
+    if all(isinstance(x, dict) for x in args):
+        output = {}
+        keys = reduce(operator.or_, [set(x) for x in args])
+
+        for key in keys:
+            # merge all of the ones that have them
+            output[key] = merge_recursive(*[x[key] for x in args if key in x])
+
+        return output
+    else:
+        return reduce(operator.add, args)
+
+def flatten(*args):
+    for x in args:
+        if hasattr(x, '__iter__'):
+            for y in flatten(*x):
+                yield y
+        else:
+            yield x
